+++ front/src/form/bsdasri/Emitter.tsx
@@ -24,6 +24,7 @@ import { useParams } from "react-router-dom";
  *
  * Emitter component with widget to group dasris
  */
+
 export function RegroupingEmitter({ status, stepName }) {
   return (
     <BaseEmitter status={status} isRegrouping={true} stepName={stepName} />
@@ -39,14 +40,51 @@ export function BaseEmitter({ status, stepName, isRegrouping = false }) {
     BsdasriStatus.Sent,
     BsdasriStatus.Received,
   ].includes(status);
+
   const emissionEmphasis = stepName === "emission";
   const { values } = useFormikContext<Bsdasri>();
   const { siret } = useParams<{ siret: string }>();
   const isUserCurrentEmitter = values?.emitter?.company?.siret === siret;
+
   return (
     <>
       {emissionEmphasis && <FillFieldsInfo />}
       {disabled && <DisabledFieldsInfo />}
+
+      {disabled && (
+        <div className="notification notification--error">
+          Les champs grisés ci-dessous ont été scellés via signature et ne sont
+          plus modifiables.
+        </div>
+      )}
+
+      {isRegrouping && (
+        <>
+          <h3 className="form__section-heading">
+            Bordereau dasri de groupement
+          </h3>
+
+          {values?.emitter?.company?.siret && !isUserCurrentEmitter && (
+            <p className="notification notification--error">
+              Pour préparer un bordereau de regroupement, vous devez y figurer
+              comme producteur
+            </p>
+          )}
+        </>
+      )}
+
+      <div
+        className={classNames("form__row", {
+          "field-emphasis": emissionEmphasis,
+        })}
+      >
+        <CompanySelector
+          disabled={disabled}
+          name="emitter.company"
+          heading="Personne responsable de l'élimination des déchets"
+          optionalMail={true}
+        />
+      </div>
 
       {disabled && (
         <div className="notification notification--error">
@@ -106,6 +144,7 @@ export function BaseEmitter({ status, stepName, isRegrouping = false }) {
         <RedErrorMessage name="emitter.onBehalfOfEcoorganisme" />
       </div>
       <h4 className="form__section-heading">Détail du déchet</h4>
+
       <div
         className={classNames("form__row", {
           "field-emphasis": emissionEmphasis,